Revert "refactor: add computer repository and save method"
